reorganize firebase files
