+++ frontend/src/config/firebase.ts
